backend/django/requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
boto3==1.34.44
botocore==1.34.45
certifi==2024.2.2
cffi==1.15.1
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==15.1.0
cryptography==41.0.2
daphne==4.0.0
Django==4.2.11
django-cors-headers==4.3.1
django-environ==0.10.0
django-otp==1.5.0
django-redis==5.4.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jmespath==1.0.1
msgpack==1.0.5
packaging==24.0
protobuf==4.21.9
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==1.7.1
pyOpenSSL==23.2.0
pyotp==2.9.0
pypng==0.20220715.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2024.1
pywatchman==1.4.1
qrcode==7.4.2
redis==4.6.0
requests==2.31.0
s3transfer==0.10.0
service-identity==23.1.0
setuptools==69.5.1
six==1.16.0
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.10.0
urllib3==1.26.18
zope.interface==6.2